Add files via upload
Description: in this project we start with performing exploratory data analysis on Sonar Dataset, polar plots are used to examine the relationship between the Range of the Sonar and the degree the object was detected. I also used a 3D scatter plot and a point cloud to perform positional analysis of the detected objects. 
Further into the project, I implemented monocular depth estimation and stereo estimation using MiDas model and the Stereo SGBM model respectively, I also created a depth map to visualise the depth estimation of every pixel.